initial commit - split tetris.tsx to several files to make it comprehensible to understand and integrate each feature from VanillaGame.tsx
