+++ src/components/rhythmia/tetris/components/index.ts
@@ -1,0 +1,15 @@
+// Re-export all components for easy importing
+export { Board } from './Board';
+export { NextPiece, HoldPiece } from './PiecePreview';
+export {
+    TitleScreen,
+    WorldDisplay,
+    ScoreDisplay,
+    ComboDisplay,
+    EnemyBar,
+    BeatBar,
+    StatsPanel,
+    ThemeNav,
+    JudgmentDisplay,
+    TouchControls,
+} from './GameUI';
