+++ src/components/rhythmia/tetris/hooks/index.ts
@@ -1,0 +1,3 @@
+// Re-export all hooks for easy importing
+export { useAudio } from './useAudio';
+export { useGameState } from './useGameState';
