+++ src/components/rhythmia/tetris/hooks/useGameState.ts
@@ -1,0 +1,282 @@
+import { useState, useRef, useEffect, useCallback } from 'react';
+import type { Piece, Board, KeyState } from '../types';
+import { BOARD_WIDTH, DEFAULT_DAS, DEFAULT_ARR, DEFAULT_SDF, ColorTheme } from '../constants';
+import { createEmptyBoard, shuffleBag, getShape, isValidPosition, createSpawnPiece } from '../utils/boardUtils';
+
+/**
+ * Custom hook for managing game state with synchronized refs
+ */
+export function useGameState() {
+    // Core game state
+    const [board, setBoard] = useState<Board>(createEmptyBoard());
+    const [currentPiece, setCurrentPiece] = useState<Piece | null>(null);
+    const [nextPiece, setNextPiece] = useState<string>('');
+    const [holdPiece, setHoldPiece] = useState<string | null>(null);
+    const [canHold, setCanHold] = useState(true);
+    const [pieceBag, setPieceBag] = useState<string[]>(shuffleBag());
+    const [score, setScore] = useState(0);
+    const [combo, setCombo] = useState(0);
+    const [lines, setLines] = useState(0);
+    const [level, setLevel] = useState(1);
+    const [gameOver, setGameOver] = useState(false);
+    const [isPaused, setIsPaused] = useState(false);
+    const [isPlaying, setIsPlaying] = useState(false);
+
+    // Rhythm game state
+    const [worldIdx, setWorldIdx] = useState(0);
+    const [enemyHP, setEnemyHP] = useState(100);
+    const [beatPhase, setBeatPhase] = useState(0);
+    const [judgmentText, setJudgmentText] = useState('');
+    const [judgmentColor, setJudgmentColor] = useState('');
+    const [showJudgmentAnim, setShowJudgmentAnim] = useState(false);
+    const [boardBeat, setBoardBeat] = useState(false);
+    const [boardShake, setBoardShake] = useState(false);
+    const [scorePop, setScorePop] = useState(false);
+
+    // DAS/ARR/SDF settings (adjustable)
+    const [das, setDas] = useState(DEFAULT_DAS);
+    const [arr, setArr] = useState(DEFAULT_ARR);
+    const [sdf, setSdf] = useState(DEFAULT_SDF);
+
+    // Color theme
+    const [colorTheme, setColorTheme] = useState<ColorTheme>('stage');
+
+    // Refs for accessing current values in callbacks (avoids stale closures)
+    const gameLoopRef = useRef<number | null>(null);
+    const beatTimerRef = useRef<number | null>(null);
+    const lastBeatRef = useRef(Date.now());
+    const lastGravityRef = useRef<number>(0);
+
+    // State refs for use in game loop
+    const boardRef = useRef<Board>(createEmptyBoard());
+    const currentPieceRef = useRef<Piece | null>(null);
+    const nextPieceRef = useRef(nextPiece);
+    const pieceBagRef = useRef<string[]>(pieceBag);
+    const holdPieceRef = useRef<string | null>(holdPiece);
+    const canHoldRef = useRef(canHold);
+    const scoreRef = useRef(0);
+    const comboRef = useRef(combo);
+    const linesRef = useRef(lines);
+    const dasRef = useRef(das);
+    const arrRef = useRef(arr);
+    const sdfRef = useRef(sdf);
+    const levelRef = useRef(level);
+    const gameOverRef = useRef(gameOver);
+    const isPausedRef = useRef(isPaused);
+    const worldIdxRef = useRef(worldIdx);
+    const enemyHPRef = useRef(enemyHP);
+    const beatPhaseRef = useRef(beatPhase);
+
+    // Key states for DAS/ARR
+    const keyStatesRef = useRef<Record<string, KeyState>>({
+        left: { pressed: false, dasCharged: false, lastMoveTime: 0, pressTime: 0 },
+        right: { pressed: false, dasCharged: false, lastMoveTime: 0, pressTime: 0 },
+        down: { pressed: false, dasCharged: false, lastMoveTime: 0, pressTime: 0 },
+    });
+
+    // Keep refs in sync with state
+    useEffect(() => { boardRef.current = board; }, [board]);
+    useEffect(() => { currentPieceRef.current = currentPiece; }, [currentPiece]);
+    useEffect(() => { nextPieceRef.current = nextPiece; }, [nextPiece]);
+    useEffect(() => { pieceBagRef.current = pieceBag; }, [pieceBag]);
+    useEffect(() => { holdPieceRef.current = holdPiece; }, [holdPiece]);
+    useEffect(() => { canHoldRef.current = canHold; }, [canHold]);
+    useEffect(() => { scoreRef.current = score; }, [score]);
+    useEffect(() => { comboRef.current = combo; }, [combo]);
+    useEffect(() => { linesRef.current = lines; }, [lines]);
+    useEffect(() => { dasRef.current = das; }, [das]);
+    useEffect(() => { arrRef.current = arr; }, [arr]);
+    useEffect(() => { sdfRef.current = sdf; }, [sdf]);
+    useEffect(() => { levelRef.current = level; }, [level]);
+    useEffect(() => { gameOverRef.current = gameOver; }, [gameOver]);
+    useEffect(() => { isPausedRef.current = isPaused; }, [isPaused]);
+    useEffect(() => { worldIdxRef.current = worldIdx; }, [worldIdx]);
+    useEffect(() => { enemyHPRef.current = enemyHP; }, [enemyHP]);
+    useEffect(() => { beatPhaseRef.current = beatPhase; }, [beatPhase]);
+
+    // Get next piece from seven-bag system
+    const getNextFromBag = useCallback((): string => {
+        let bag = [...pieceBagRef.current];
+        if (bag.length === 0) {
+            bag = shuffleBag();
+        }
+        const piece = bag.shift()!;
+        setPieceBag(bag);
+        return piece;
+    }, []);
+
+    // Spawn a new piece
+    const spawnPiece = useCallback((): Piece | null => {
+        const type = nextPieceRef.current;
+        const newPiece = createSpawnPiece(type);
+
+        setNextPiece(getNextFromBag());
+        setCanHold(true);
+
+        if (!isValidPosition(newPiece, boardRef.current)) {
+            setGameOver(true);
+            setIsPlaying(false);
+            return null;
+        }
+
+        return newPiece;
+    }, [getNextFromBag]);
+
+    // Show judgment text with animation
+    const showJudgment = useCallback((text: string, color: string) => {
+        setJudgmentText(text);
+        setJudgmentColor(color);
+        setShowJudgmentAnim(false);
+        requestAnimationFrame(() => {
+            setShowJudgmentAnim(true);
+        });
+    }, []);
+
+    // Update score with pop animation
+    const updateScore = useCallback((newScore: number) => {
+        setScore(newScore);
+        setScorePop(true);
+        setTimeout(() => setScorePop(false), 100);
+    }, []);
+
+    // Trigger board shake effect
+    const triggerBoardShake = useCallback(() => {
+        setBoardShake(true);
+        setTimeout(() => setBoardShake(false), 200);
+    }, []);
+
+    // Reset key states
+    const resetKeyStates = useCallback(() => {
+        keyStatesRef.current = {
+            left: { pressed: false, dasCharged: false, lastMoveTime: 0, pressTime: 0 },
+            right: { pressed: false, dasCharged: false, lastMoveTime: 0, pressTime: 0 },
+            down: { pressed: false, dasCharged: false, lastMoveTime: 0, pressTime: 0 },
+        };
+    }, []);
+
+    // Initialize/reset game
+    const initGame = useCallback(() => {
+        setBoard(createEmptyBoard());
+        boardRef.current = createEmptyBoard();
+        setScore(0);
+        setCombo(0);
+        setLines(0);
+        setLevel(1);
+        setWorldIdx(0);
+        setEnemyHP(100);
+        setGameOver(false);
+        setIsPaused(false);
+        setIsPlaying(true);
+        setHoldPiece(null);
+        setCanHold(true);
+
+        resetKeyStates();
+
+        // Initialize seven-bag system
+        const bag = shuffleBag();
+        const type = bag[0];
+        const next = bag[1];
+        setPieceBag(bag.slice(2));
+
+        setNextPiece(next);
+        nextPieceRef.current = next;
+
+        const shape = getShape(type, 0);
+        const initialPiece: Piece = {
+            type,
+            rotation: 0,
+            x: Math.floor((BOARD_WIDTH - shape[0].length) / 2),
+            y: type === 'I' ? -2 : -1,
+        };
+
+        setCurrentPiece(initialPiece);
+        currentPieceRef.current = initialPiece;
+        lastGravityRef.current = performance.now();
+    }, [resetKeyStates]);
+
+    return {
+        // State
+        board,
+        currentPiece,
+        nextPiece,
+        holdPiece,
+        canHold,
+        pieceBag,
+        score,
+        combo,
+        lines,
+        level,
+        gameOver,
+        isPaused,
+        isPlaying,
+        worldIdx,
+        enemyHP,
+        beatPhase,
+        judgmentText,
+        judgmentColor,
+        showJudgmentAnim,
+        boardBeat,
+        boardShake,
+        scorePop,
+        das,
+        arr,
+        sdf,
+        colorTheme,
+
+        // Setters
+        setBoard,
+        setCurrentPiece,
+        setNextPiece,
+        setHoldPiece,
+        setCanHold,
+        setPieceBag,
+        setScore,
+        setCombo,
+        setLines,
+        setLevel,
+        setGameOver,
+        setIsPaused,
+        setIsPlaying,
+        setWorldIdx,
+        setEnemyHP,
+        setBeatPhase,
+        setBoardBeat,
+        setDas,
+        setArr,
+        setSdf,
+        setColorTheme,
+
+        // Refs
+        boardRef,
+        currentPieceRef,
+        nextPieceRef,
+        pieceBagRef,
+        holdPieceRef,
+        canHoldRef,
+        scoreRef,
+        comboRef,
+        linesRef,
+        dasRef,
+        arrRef,
+        sdfRef,
+        levelRef,
+        gameOverRef,
+        isPausedRef,
+        worldIdxRef,
+        enemyHPRef,
+        beatPhaseRef,
+        keyStatesRef,
+        gameLoopRef,
+        beatTimerRef,
+        lastBeatRef,
+        lastGravityRef,
+
+        // Actions
+        spawnPiece,
+        showJudgment,
+        updateScore,
+        triggerBoardShake,
+        resetKeyStates,
+        initGame,
+    };
+}
