+++ src/components/rhythmia/tetris/index.ts
@@ -1,0 +1,7 @@
+// Main Tetris/Rhythmia game component
+export { default as Rhythmia } from './tetris';
+export { default } from './tetris';
+
+// Re-export types and constants for external use
+export * from './types';
+export * from './constants';
