+++ src/components/rhythmia/tetris/utils/index.ts
@@ -1,0 +1,2 @@
+// Re-export all utilities for easy importing
+export * from './boardUtils';
